Add support for azure-storage-python 12
