+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py{27,34,35,36}
+envlist = py{27,34,35,36},py36-old-azure-storage-blob
 
 [pytest]
 pep8ignore = test_* E402
@@ -16,7 +16,6 @@ deps=
   pytest-mock
   pytest-xdist
   mock
-  tempdir
   redis
   psycopg2
   sqlalchemy
@@ -27,4 +26,14 @@ deps=
   azure-storage-blob
   futures
 # ideally we would not need futures here but it doesn't work otherwise
-commands=py.test -n 4 --dist=loadfile --cov=simplekv -rs --pep8 --doctest-modules simplekv/idgen.py simplekv/fs.py tests
+commands=pytest -n 4 --dist=loadfile --cov=simplekv -rs --pep8 --doctest-modules simplekv/idgen.py simplekv/fs.py tests
+
+[testenv:py36-old-azure-storage-blob]
+# Test on one python version with the "old" code for azure-storage-blob<12
+deps=
+  pytest
+  pytest-cov
+  pytest-mock
+  mock
+  azure-storage-blob<12
+commands=pytest tests/test_azure_store.py
